[SETUP] 0x02. Minimum Operations CREATE Directory 0x02-Minimum_Operations CREATE README.md CREATE image for 0x02 Minimum Operations
Tasks
0. Minimum Operations

In a text file, there is a single character H. Your text editor can execute only two operations in this file: Copy All and Paste. Given a number n, write a method that calculates the fewest number of operations needed to result in exactly n H characters in the file.

Prototype: def minOperations(n)
Returns an integer
If n is impossible to achieve, return 0